[cegar-ocp] print number of concrete states in subtask

+++ src/search/utils/math.cc
@@ -3,7 +3,8 @@
 #include <cassert>
 
 namespace utils {
-bool is_product_within_limit(int factor1, int factor2, int limit) {
+bool is_product_within_limit(
+    long long factor1, long long factor2, long long limit) {
     assert(factor1 >= 0);
     assert(factor2 >= 0);
     assert(limit >= 0);
